Move static method matcher pointcut to the correct namespace #1
