+++ src/Go/Aop/Pointcut/AnnotationMethodPointcut.php
@@ -11,7 +11,6 @@ namespace Go\Aop\Pointcut;
 use ReflectionMethod;
 
 use Go\Instrument\RawAnnotationReader;
-use Go\Aop\Support\StaticMethodMatcherPointcut;
 
 use TokenReflection\ReflectionMethod as ParsedReflectionMethod;
