+++ src/Go/Aop/Pointcut/SignatureMethodPointcut.php
@@ -11,7 +11,6 @@ namespace Go\Aop\Pointcut;
 use ReflectionMethod;
 
 use Go\Aop\PointFilter;
-use Go\Aop\Support\StaticMethodMatcherPointcut;
 
 use TokenReflection\ReflectionMethod as ParsedReflectionMethod;
